requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
py-spy==0.4.0
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
statsig_python_core==0.5.0
typing_extensions==4.14.0
urllib3==2.5.0
